% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-cmp.R
\name{hash_compare}
\alias{hash_compare}
\title{Compare two hashes}
\usage{
hash_compare(sig1, sig2)
}
\arguments{
\item{sig1, sig2}{character signatures created with one of the hash generation functions}
}
\value{
integer(1)
}
\description{
Compare two hashes
}
\examples{
system.file("extdat", package="ssdeepr") \%>\%
list.files(full.names = TRUE, pattern = "html$", include.dirs = FALSE) \%>\%
  hash_file() -> hashes

hash_compare(hashes$hash[1], hashes$hash[1])
hash_compare(hashes$hash[1], hashes$hash[2])
hash_compare(hashes$hash[1], hashes$hash[3])
hash_compare(hashes$hash[1], hashes$hash[4])
}
