% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-file.R
\name{hash_file}
\alias{hash_file}
\title{Return CTP hash of one or more files}
\usage{
hash_file(path)
}
\arguments{
\item{path}{character vector of file paths}
}
\value{
data frame
}
\description{
Return CTP hash of one or more files
}
\examples{
system.file("extdat", package="ssdeepr") \%>\%
list.files(full.names = TRUE, pattern = "html$", include.dirs = FALSE) \%>\%
  hash_file() -> hashes

hash_compare(hashes$hash[1], hashes$hash[1])
hash_compare(hashes$hash[1], hashes$hash[2])
hash_compare(hashes$hash[1], hashes$hash[3])
hash_compare(hashes$hash[1], hashes$hash[4])
}
