% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-buf.R
\name{hash_raw}
\alias{hash_raw}
\title{Return CTP hash of a raw vector}
\usage{
hash_raw(x)
}
\arguments{
\item{x}{raw vector or length 1 character vector}
}
\value{
character(1)
}
\description{
Return CTP hash of a raw vector
}
\note{
Length 1 character vectors will be auto-converted to a raw vectors
}
\examples{
ltrs <- paste0(c(letters, LETTERS), collapse="")
hash_raw(ltrs)
}
