% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-hash.R
\name{split_hash}
\alias{split_hash}
\title{Splits an ssdeep hash string into component parts}
\usage{
split_hash(x)
}
\arguments{
\item{x}{character vector of ssdeep hashes}
}
\value{
data frame
}
\description{
ssdeep hashes consists of three parts, separated by \code{:}:
\itemize{
\item chunk size : size of the chunks in rest of the hash
\item chunk : where each character represents a part of the file of length "chunk size"
\item double chunk : hash computation where "chunk size" = 2 * "chunk size"
This function splits the single has string into three component parts.
}
}
\examples{
split_hash(hash_con(url("https://en.wikipedia.org/wiki/Donald_Knuth")))
}
