% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdeepr-package.R
\docType{package}
\name{ssdeepr}
\alias{ssdeepr}
\alias{ssdeepr-package}
\title{Context Triggered Piecewise Hash Computation Using 'ssdeep'}
\description{
The ssdeep project provides an open source library \url{https://github.com/ssdeep-project/ssdeep/}
context triggered piecewise hashing. Methods are provided to compute and compare
hashes from character/byte streams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/ssdeepr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/ssdeepr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
