% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-con.R
\name{hash_con}
\alias{hash_con}
\title{Return CTP hash of one data collected from a connection}
\usage{
hash_con(con)
}
\arguments{
\item{con}{a \link{connection}}
}
\value{
data frame
}
\description{
Return CTP hash of one data collected from a connection
}
\examples{
hash_con(url("https://en.wikipedia.org/wiki/Donald_Knuth"))
}
