% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdeepr-package.R
\docType{package}
\name{ssdeepr}
\alias{ssdeepr}
\alias{ssdeepr-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/ssdeepr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/ssdeepr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
